Add robolectric, use new pluralStringResource, some clean up (#28)
* Finished PlayerViewModel tests

* upgrade android version, fix warnings

* upgrade plugin

* upgrade mockk

* use new M3 text field

* use plurasl, robolectric

* deleted unused file

+++ app/src/main/java/com/sebastianvm/musicplayer/database/daos/AlbumDao.kt
@@ -11,7 +11,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface AlbumDao {
     @Query("SELECT COUNT(*) FROM Album")
-    fun getAlbumsCount(): Flow<Long>
+    fun getAlbumsCount(): Flow<Int>
 
     @Transaction
     @Query("SELECT * from Album WHERE Album.albumId IN (:albumIds)")
