+++ app/src/main/java/com/sebastianvm/musicplayer/database/daos/GenreDao.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface GenreDao {
     @Query("SELECT COUNT(*) FROM Genre")
-    fun getGenresCount(): Flow<Long>
+    fun getGenresCount(): Flow<Int>
 
     @Query("SELECT * FROM Genre")
     fun getGenres(): Flow<List<Genre>>
