+++ app/src/main/java/com/sebastianvm/musicplayer/repository/genre/GenreRepository.kt
@@ -4,6 +4,6 @@ import com.sebastianvm.musicplayer.database.entities.Genre
 import kotlinx.coroutines.flow.Flow
 
 interface GenreRepository {
-    fun getGenresCount(): Flow<Long>
+    fun getGenresCount(): Flow<Int>
     fun getGenres(): Flow<List<Genre>>
 }
